add pages support
