+++ package.json
@@ -13,6 +13,7 @@
 		"react": "^19.2.5",
 		"react-colorful": "^5.6.1",
 		"react-dom": "^19.2.5",
+		"react-router-dom": "^7.1.5",
 		"zod": "^4.3.6"
 	},
 	"devDependencies": {
